Safari-609.1.11.1/LayoutTests/fast/events/beforeunload-alert-expected.txt

CONSOLE MESSAGE: line 8: Use of window.alert is not allowed while unloading a page.
This test verifies that prompts are forbidden when leaving a webpage.

Prompts include alert, confirm, prompt, and showModalDialog.

If there is no prompt, the test passed.
